extracts collaborating teams component and fixes minor styles issues
