+++ packages/react-components/src/organisms/ProjectContributors.tsx
@@ -1,22 +1,16 @@
 import { CollaboratingTeam, FundedTeam, ProjectMember } from '@asap-hub/model';
-import { network, sharedResearch } from '@asap-hub/routing';
 import { css } from '@emotion/react';
 import { useState } from 'react';
 
-import { Button, Card, Headline3, Link, Pill, TabButton } from '../atoms';
+import { Card, Headline3, Pill, TabButton } from '../atoms';
 import { lead, steel } from '../colors';
 import {
-  article as articleIcon,
-  DiscoveryTeamIcon,
-  ResourceTeamIcon,
-  InactiveBadgeIcon,
-  chevronUpIcon,
-  chevronDownIcon,
-} from '../icons';
-import { TabNav, ProjectMembers, LinkHeadline } from '../molecules';
-import { rem, tabletScreen } from '../pixels';
-
-const nonMobileQuery = `@media (min-width: ${tabletScreen.min}px)`;
+  TabNav,
+  ProjectMembers,
+  LinkHeadline,
+  CollaboratingTeams,
+} from '../molecules';
+import { rem } from '../pixels';
 
 const cardContentStyles = css({
   padding: `${rem(32)} ${rem(24)}`,
@@ -51,143 +45,6 @@ const teamDescriptionStyles = css({
   margin: 0,
 });
 
-const collaboratorsListStyles = css({
-  display: 'flex',
-  flexDirection: 'column',
-});
-
-const collaboratorRowStyles = css({
-  borderTop: `1px solid ${steel.rgb}`,
-  padding: `${rem(16)} 0`,
-  ':first-of-type': {
-    borderTop: 'none',
-    paddingTop: 8,
-  },
-});
-
-const collaboratorHeaderStyles = css({
-  display: 'flex',
-  // Mobile: chevron aligns to the bottom line (the article count).
-  alignItems: 'flex-end',
-  justifyContent: 'space-between',
-  gap: rem(12),
-  cursor: 'pointer',
-  [nonMobileQuery]: {
-    alignItems: 'center',
-  },
-});
-
-const chevronButtonStyles = css({
-  background: 'none',
-  border: 'none',
-  cursor: 'pointer',
-  padding: 0,
-  display: 'inline-flex',
-  alignItems: 'center',
-});
-
-const collaboratorHeaderLeftStyles = css({
-  display: 'flex',
-  // Mobile: team name on the first line, article count on the next.
-  flexDirection: 'column',
-  alignItems: 'flex-start',
-  gap: rem(8),
-  [nonMobileQuery]: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    flexWrap: 'wrap',
-  },
-});
-
-const teamNameGroupStyles = css({
-  display: 'inline-flex',
-  alignItems: 'center',
-  gap: rem(8),
-});
-
-const collaboratorNameStyles = css({
-  display: 'inline-flex',
-  alignItems: 'center',
-  gap: rem(6),
-});
-
-const articleCountStyles = css({
-  color: lead.rgb,
-  paddingLeft: rem(16),
-  '& > [data-bullet]': {
-    display: 'none',
-  },
-  [nonMobileQuery]: {
-    '& > [data-bullet]': {
-      display: 'inline',
-    },
-  },
-});
-
-const ARTICLES_BEFORE_SCROLL = 7;
-const ARTICLE_ROW_HEIGHT = 40;
-
-const articlesListStyles = (count: number) =>
-  css({
-    listStyle: 'none',
-    margin: 0,
-    padding: `${rem(12)} 0 ${rem(20)} ${rem(32)}`,
-    display: 'flex',
-    flexDirection: 'column',
-    gap: rem(8),
-    ...(count > ARTICLES_BEFORE_SCROLL
-      ? {
-          maxHeight: rem(ARTICLES_BEFORE_SCROLL * ARTICLE_ROW_HEIGHT),
-          overflowY: 'auto',
-        }
-      : {}),
-  });
-
-const articleRowStyles = css({
-  // Mobile: title row on top, pill flush-left underneath.
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'flex-start',
-  gap: rem(8),
-  [nonMobileQuery]: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-});
-
-const articleTitleGroupStyles = css({
-  display: 'inline-flex',
-  alignItems: 'center',
-  gap: rem(8),
-});
-
-const emptyStateStyles = css({
-  color: lead.rgb,
-  margin: 0,
-  padding: `${rem(24)} 0`,
-  textAlign: 'left',
-});
-
-const viewMoreContainerStyles = css({
-  display: 'flex',
-  justifyContent: 'center',
-  marginTop: rem(16),
-  paddingTop: rem(14),
-  // Negative horizontal margins cancel the parent card's 24px horizontal
-  // padding so the divider spans the full card width.
-  marginInline: `-${rem(24)}`,
-  marginBottom: `-${rem(14)}`,
-  borderTop: `1px solid ${steel.rgb}`,
-});
-
-const INITIAL_COLLABORATORS_COUNT = 10;
-
-const articleTypeLabel = (type?: string) => {
-  if (!type) return undefined;
-  if (type === 'Published') return 'Publication';
-  return type;
-};
-
 type ProjectContributorsProps =
   | {
       fundedTeam: FundedTeam;
@@ -211,96 +68,6 @@ export const memberTabs = ['Project Members'] as const;
 export type Tabs = (typeof tabs)[number];
 export type MemberTabs = (typeof memberTabs)[number];
 
-const teamTypeIconFor = (teamType?: string): React.FC =>
-  teamType === 'Resource Team' ? ResourceTeamIcon : DiscoveryTeamIcon;
-
-const CollaboratingTeamRow: React.FC<{
-  team: CollaboratingTeam;
-}> = ({ team }) => {
-  const TeamTypeIcon = teamTypeIconFor(team.teamType);
-  const [expanded, setExpanded] = useState(false);
-  const articleCount = team.articles.length;
-
-  const toggle = () => setExpanded((v) => !v);
-
-  return (
-    <div css={collaboratorRowStyles}>
-      <div
-        css={collaboratorHeaderStyles}
-        role="button"
-        tabIndex={0}
-        aria-expanded={expanded}
-        aria-label={
-          expanded
-            ? `Collapse ${team.displayName} articles`
-            : `Expand ${team.displayName} articles`
-        }
-        onClick={toggle}
-        onKeyDown={(e) => {
-          if (e.key === 'Enter' || e.key === ' ') {
-            e.preventDefault();
-            toggle();
-          }
-        }}
-      >
-        <span css={collaboratorHeaderLeftStyles}>
-          <span css={teamNameGroupStyles}>
-            <span aria-hidden="true">
-              <TeamTypeIcon />
-            </span>
-            <span
-              css={collaboratorNameStyles}
-              onClick={(e) => e.stopPropagation()}
-            >
-              <Link href={network({}).teams({}).team({ teamId: team.id }).$}>
-                {team.displayName}
-              </Link>
-              {team.inactiveSince && <InactiveBadgeIcon />}
-            </span>
-          </span>
-          <span css={articleCountStyles}>
-            <span data-bullet aria-hidden="true">
-              {'• '}
-            </span>
-            {articleCount} {articleCount === 1 ? 'Article' : 'Articles'}
-          </span>
-        </span>
-        <span css={chevronButtonStyles} aria-hidden="true">
-          {expanded ? chevronUpIcon : chevronDownIcon}
-        </span>
-      </div>
-      {expanded && (
-        <ul css={articlesListStyles(articleCount)}>
-          {team.articles.map((article) => {
-            const label = articleTypeLabel(article.type);
-            return (
-              <li key={article.id} css={articleRowStyles}>
-                <span css={articleTitleGroupStyles}>
-                  <span aria-hidden="true">{articleIcon}</span>
-                  <Link
-                    href={
-                      sharedResearch({}).researchOutput({
-                        researchOutputId: article.id,
-                      }).$
-                    }
-                  >
-                    {article.title}
-                  </Link>
-                </span>
-                {label && (
-                  <Pill noMargin accent="gray">
-                    {label}
-                  </Pill>
-                )}
-              </li>
-            );
-          })}
-        </ul>
-      )}
-    </div>
-  );
-};
-
 const ProjectContributors: React.FC<ProjectContributorsProps> = ({
   fundedTeam,
   collaboratingTeams,
@@ -308,7 +75,6 @@ const ProjectContributors: React.FC<ProjectContributorsProps> = ({
   showTeamInfo = false,
 }) => {
   const [activeTab, setActiveTab] = useState<Tabs>('Funded Team');
-  const [showAllCollaborators, setShowAllCollaborators] = useState(false);
 
   if (projectMembers) {
     // Trainee or Resource individual-based projects - show Project Members only
@@ -337,11 +103,6 @@ const ProjectContributors: React.FC<ProjectContributorsProps> = ({
   }
 
   const collaboratorsCount = collaboratingTeams?.length ?? 0;
-  const visibleCollaborators =
-    showAllCollaborators || collaboratorsCount <= INITIAL_COLLABORATORS_COUNT
-      ? collaboratingTeams ?? []
-      : (collaboratingTeams ?? []).slice(0, INITIAL_COLLABORATORS_COUNT);
-  const hasMoreCollaborators = collaboratorsCount > INITIAL_COLLABORATORS_COUNT;
 
   return (
     <Card padding={false}>
@@ -383,30 +144,8 @@ const ProjectContributors: React.FC<ProjectContributorsProps> = ({
               </LinkHeadline>
               <p css={teamDescriptionStyles}>{fundedTeam.teamDescription}</p>
             </div>
-          ) : collaboratorsCount === 0 ? (
-            <p css={emptyStateStyles}>
-              There are no team collaborations on this project yet.
-            </p>
           ) : (
-            <>
-              <div css={collaboratorsListStyles}>
-                {visibleCollaborators.map((team) => (
-                  <CollaboratingTeamRow key={team.id} team={team} />
-                ))}
-              </div>
-              {hasMoreCollaborators && (
-                <div css={viewMoreContainerStyles}>
-                  <Button
-                    linkStyle
-                    onClick={() => setShowAllCollaborators((v) => !v)}
-                  >
-                    {showAllCollaborators
-                      ? 'View Less Collaborators'
-                      : 'View More Collaborators'}
-                  </Button>
-                </div>
-              )}
-            </>
+            <CollaboratingTeams collaboratingTeams={collaboratingTeams} />
           )}
         </div>
       </div>
