+++ packages/react-components/src/organisms/__tests__/ProjectContributors.test.tsx
@@ -115,7 +115,20 @@ describe('ProjectContributors', () => {
       expect(screen.getByText('Collaborators (3)')).toBeInTheDocument();
     });
 
-    it('renders empty state when there are no collaborating teams', async () => {
+    it('renders the collaborating teams in the Collaborators tab', async () => {
+      renderWithRouter(
+        <ProjectContributors
+          fundedTeam={mockFundedTeam}
+          collaboratingTeams={makeCollaboratingTeams(2)}
+        />,
+      );
+      await userEvent.click(screen.getByText('Collaborators (2)'));
+      expect(
+        screen.getByRole('button', { name: /Expand Team A articles/i }),
+      ).toBeInTheDocument();
+    });
+
+    it('renders the empty state in the Collaborators tab', async () => {
       renderWithRouter(
         <ProjectContributors
           fundedTeam={mockFundedTeam}
@@ -126,239 +139,6 @@ describe('ProjectContributors', () => {
       expect(
         screen.getByText(/no team collaborations on this project/i),
       ).toBeInTheDocument();
-    });
-
-    it('renders collaborating teams with article counts and expands to show articles', async () => {
-      const teams: CollaboratingTeam[] = [
-        {
-          id: 'team-a',
-          displayName: 'Team Alpha',
-          articles: [
-            { id: 'a-1', title: 'Alpha One', type: 'Preprint' },
-            { id: 'a-2', title: 'Alpha Two', type: 'Published' },
-          ],
-        },
-      ];
-
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={teams}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (1)'));
-      expect(screen.getByText(/2 Articles/)).toBeInTheDocument();
-
-      await userEvent.click(
-        screen.getByRole('button', { name: /Expand Team Alpha articles/i }),
-      );
-      expect(screen.getByText('Alpha One')).toBeInTheDocument();
-      expect(screen.getByText('Alpha Two')).toBeInTheDocument();
-      // 'Published' renders as 'Publication' pill
-      expect(screen.getByText('Publication')).toBeInTheDocument();
-      expect(screen.getByText('Preprint')).toBeInTheDocument();
-    });
-
-    it('shows View More when there are more than 10 collaborating teams', async () => {
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={makeCollaboratingTeams(12)}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (12)'));
-
-      // Only first 10 visible initially
-      expect(
-        screen.getAllByRole('button', {
-          name: /Expand Team [A-Z] articles/,
-        }),
-      ).toHaveLength(10);
-
-      await userEvent.click(
-        screen.getByRole('button', { name: /View More Collaborators/i }),
-      );
-      expect(
-        screen.getAllByRole('button', {
-          name: /Expand Team [A-Z] articles/,
-        }),
-      ).toHaveLength(12);
-    });
-
-    it('AC02: makes the articles list scrollable when there are more than 7 articles', async () => {
-      const teams: CollaboratingTeam[] = [
-        {
-          id: 'team-many',
-          displayName: 'Team Many',
-          articles: Array.from({ length: 8 }, (_, i) => ({
-            id: `m-${i}`,
-            title: `Many ${i}`,
-            type: 'Preprint' as const,
-          })),
-        },
-      ];
-
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={teams}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (1)'));
-      await userEvent.click(
-        screen.getByRole('button', { name: /Expand Team Many articles/i }),
-      );
-
-      // The articles list is the only <ul> containing the article titles
-      const list = screen.getByText('Many 0').closest('ul');
-      expect(list).not.toBeNull();
-      expect(list as HTMLElement).toHaveStyleRule('overflow-y', 'auto');
-      expect(list as HTMLElement).toHaveStyleRule(
-        'max-height',
-        expect.stringMatching(/em$/),
-      );
-    });
-
-    it('AC03: links each article title to its research output detail page', async () => {
-      const teams: CollaboratingTeam[] = [
-        {
-          id: 'team-a',
-          displayName: 'Team Alpha',
-          articles: [{ id: 'ro-42', title: 'Some Article', type: 'Preprint' }],
-        },
-      ];
-
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={teams}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (1)'));
-      await userEvent.click(
-        screen.getByRole('button', { name: /Expand Team Alpha articles/i }),
-      );
-
-      expect(
-        screen.getByRole('link', { name: 'Some Article' }),
-      ).toHaveAttribute('href', '/shared-research/ro-42');
-    });
-
-    it('AC04: links each collaborating team name to its team detail page', async () => {
-      const teams: CollaboratingTeam[] = [
-        {
-          id: 'team-xyz',
-          displayName: 'Team XYZ',
-          articles: [{ id: 'a-1', title: 'A1', type: 'Preprint' }],
-        },
-      ];
-
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={teams}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (1)'));
-
-      expect(screen.getByRole('link', { name: 'Team XYZ' })).toHaveAttribute(
-        'href',
-        '/network/teams/team-xyz',
-      );
-    });
-
-    it('uses each collaborating team’s own type to pick its icon', async () => {
-      const teams: CollaboratingTeam[] = [
-        {
-          id: 'discovery-team',
-          displayName: 'Discovery Co',
-          teamType: 'Discovery Team',
-          articles: [{ id: 'a-1', title: 'A1', type: 'Preprint' }],
-        },
-        {
-          id: 'resource-team',
-          displayName: 'Resource Co',
-          teamType: 'Resource Team',
-          articles: [{ id: 'a-2', title: 'A2', type: 'Preprint' }],
-        },
-      ];
-
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={teams}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (2)'));
-
-      const discoveryRow = screen
-        .getByRole('link', { name: 'Discovery Co' })
-        .closest('[role="button"]') as HTMLElement;
-      const resourceRow = screen
-        .getByRole('link', { name: 'Resource Co' })
-        .closest('[role="button"]') as HTMLElement;
-
-      expect(discoveryRow.querySelector('title')?.textContent).toBe(
-        'Discovery Team Icon',
-      );
-      expect(resourceRow.querySelector('title')?.textContent).toBe(
-        'Resource Team Icon',
-      );
-    });
-
-    it('expands when clicking anywhere on the row, but not on the team link', async () => {
-      const teams: CollaboratingTeam[] = [
-        {
-          id: 'team-a',
-          displayName: 'Team Alpha',
-          articles: [{ id: 'a-1', title: 'Alpha One', type: 'Preprint' }],
-        },
-      ];
-
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={teams}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (1)'));
-
-      // Clicking the link itself does NOT toggle the row
-      await userEvent.click(screen.getByRole('link', { name: 'Team Alpha' }));
-      expect(screen.queryByText('Alpha One')).not.toBeInTheDocument();
-
-      // Clicking the article-count area (anywhere else in the row) DOES toggle
-      await userEvent.click(screen.getByText(/1 Article$/));
-      expect(screen.getByText('Alpha One')).toBeInTheDocument();
-    });
-
-    it('toggles row via Enter and Space keys', async () => {
-      const teams: CollaboratingTeam[] = [
-        {
-          id: 'team-a',
-          displayName: 'Team Alpha',
-          articles: [{ id: 'a-1', title: 'Alpha One', type: 'Preprint' }],
-        },
-      ];
-
-      renderWithRouter(
-        <ProjectContributors
-          fundedTeam={mockFundedTeam}
-          collaboratingTeams={teams}
-        />,
-      );
-      await userEvent.click(screen.getByText('Collaborators (1)'));
-
-      const row = screen.getByRole('button', {
-        name: /Expand Team Alpha articles/i,
-      });
-      row.focus();
-
-      await userEvent.keyboard('{Enter}');
-      expect(screen.getByText('Alpha One')).toBeInTheDocument();
-
-      await userEvent.keyboard(' ');
-      expect(screen.queryByText('Alpha One')).not.toBeInTheDocument();
     });
 
     it('switches back to the Funded Team tab from Collaborators', async () => {
